+++ CHANGELOG.md
@@ -18,6 +18,7 @@ A `secondaryLevel: true` option is available to add operations to the widget's c
 * Fixes `range` field type default value not being set properly.
 * Fixes autocomplete and search sorting and as a consequence, fixes potential duplicates during pagination.
 * Fixes all eslint warnings.
+* When pasting a widget from the clipboard, the correct widget type is always offered on the "Add Content" menu.
 * Widget live preview is now attempting to auto-position the Widget Editor modal only if no explicit widget configuration (`options.origin`) is provided.
 
 ### Changes